start moving training data stuff into new directory
