+++ src/mcts/node.h
@@ -38,7 +38,7 @@
 #include "chess/position.h"
 #include "neural/cache.h"
 #include "neural/encoder.h"
-#include "neural/writer.h"
+#include "trainingdata/writer.h"
 #include "proto/net.pb.h"
 #include "utils/mutex.h"
