+++ src/selfplay/game.cc
@@ -32,7 +32,7 @@
 #include "mcts/stoppers/common.h"
 #include "mcts/stoppers/factory.h"
 #include "mcts/stoppers/stoppers.h"
-#include "neural/writer.h"
+#include "trainingdata/writer.h"
 
 namespace lczero {
