+++ src/trainingdata/reader.cc
@@ -25,7 +25,7 @@
   Program grant you additional permission to convey the resulting work.
 */
 
-#include "neural/reader.h"
+#include "trainingdata/reader.h"
 
 #include "neural/encoder.h"
 #include "utils/bititer.h"
