+++ src/trainingdata/writer.h
